Removed redundant Logger classes, no need.
Fixed issue in AccessTools that would not access private fields when setting.

Created a ConsoleController class that deals with allocation and freeing of the console. Deals with subscribing to the Application logging system.

Fixed GetRandom not using (this) modifier.

Added IsDebug property to the Info struct.

Mod's now introduce the assembly that the mod was loaded from, along with a function to Update a settings file.

Overhauled the ModAggreation functions. Many were redundant and caused O(n^2) times because of reiterations.

Instead CreateMods is called then each mod is Initialized. This is different because when a Mod folder is parsed it created an AggregatedMod objec that includes more information about the collected mod.